requirements.txt
Flask
flask-cors
python-dotenv
google-generative-ai
gunicorn
